lib/libc/amd64: add archlevel-based simd dispatch framework (DRAFT)
Summary:
Add a framework for selecting from one of multiple implementations
of a function based on amd64 architecture level (cf. amd64 SysV
ABI supplement).  As a sample use, a set of optimised strlen
functions is implemented.

Documentation is provided in the new manual page simd(7).

This DR is a draft of the dispatch logic to get some feedback on
whether the general design and concepts are sound.

Open questions:

 - Should we provide a way to override the detected architecture
   level from within a program using libc?  This could e.g. be
   done by defining a weak symbol that can be overridden to supply
   a custom archlevel limit.
 - I haven't done a whole lot of benchmarking on this; kib@ mentioned
   that using SIMD inside the libc may lead to worse performance for
   applications that usually do not use SIMD due to lazy state
   initialisation.  Unclear if this still applies as the C compiler
   has started to use SIMD instructions in ordinary code anyway.
   Also looking for recommendations for benchmarks to use for this
   work (apart from writing custom microbenchmarks for the individual
   functions)
 - Should we provide a utility to print the current architecture level?

Sponsored by:	FreeBSD Foundation

Test Plan:
libc test suite passes.  Test suite needs to be expanded so it tests all
architecture levels of each SIMD-enhanced function that can be run on
the test machine.  Benchmarks need to be written, too.

Reviewers: mjg

Subscribers: jrm, imp, kib

Differential Revision: https://reviews.freebsd.org/D40693