lib/libc/tests/string/memcmp_test.c: extend test to support custom memcmp function
Summary:
Extend the tests to permit loading an external memcmp function
and testing it over using the libc version. This was added by the
example of other tests in the test suite doing the same thing and
helped tremendously in development.

This change was originally part of D41442 but was taken out to
permit separate review as extrapolated from @ngie's request in
D41349.

Sponsored by:	FreeBSD Foundation

Test Plan: n/a

Reviewers: ngie

Subscribers: ngie, imp, mjg

Differential Revision: https://reviews.freebsd.org/D41528