libexec/rtld-elf/rtld-libc/Makefile.inc: do not use machdep string functions
Summary:
With ifunc based SIMD dispatch, we cannot use the amd64 assembly
implementations of string functions.  Modify rtld to instead use
the generic functions.  To avoid an architecture-specific special
case, this change is applied to all architectures.

This change is a prerequisite to and formerly part of D40693.

Sponsored by:	FreeBSD Foundation

Test Plan: Passed `make tinderbox` as part of D40693.

Reviewers: kib

Subscribers: imp, mjg, bdragon

Differential Revision: https://reviews.freebsd.org/D41050