lib/libc/amd64: back out amd64 SIMD work
Summary:
Due to some bugs being found, back out the amd64 SIMD libc work from
releng/14.0 ahead of the release and give these changes in a critical
part of FreeBSD some more time to mature. This was requested by @imp
and @kevans.

This patch set backs out D40693, D41333, D41349, D41442, and D41697.
Change sets D40729, D40730 (ffs), D41011 (bcmp integer overflow bug),
and D41050 (rtld bits) are not backed out.  The intent is for this
backout to be a direct commit to releng/14.0 with the changes staying on
CURRENT and stable/14.

Sponsored by:	The FreeBSD Foundation

Test Plan: passes a build test

Reviewers: imp, kevans, gjb

Subscribers: imp, kevans, jrm, kib, mjg

Differential Revision: https://reviews.freebsd.org/D41855