contrib/netbsd-tests/lib/libc/string/t_strcmp.c: extend and add support for external strcmp function
Summary:
This changeset adds support for testing an external strcmp function
as is already the case for some other unit tests.
This greatly simplifies the testing of borked strcmp() implementations
that do not perform well enough to even start up ATK correctly.

We also add a new set of tests that comprehensively test strcmp() on
various alignments of the input.  This made it easy to smoke out many
exciting new bugs in my upcoming SSE strcmp() implementation.

Sponsored by:	The FreeBSD Foundation

Test Plan: n/a

Reviewers: ngie

Subscribers: imp, mjg

Differential Revision: https://reviews.freebsd.org/D41970