lib/libc/amd64/string/stpncpy.S: fix possible integer overflow (-0.08%)
The original code compared rdx+rcx with 32, where rdx is the
buffer length and 0<=rcx<16 is the alignment offset.  This
comparison can overflow if rdx is close to SIZE_MAX.  While
such buffer lengths are not reasonable, the code should still
behave predictably and crash when accessing unmapped storage
rather than going into the wrong case due to overflow.

Fix the issue by instead comparing rdx with 32-rcx, which is
not subject to overflow.  To allow us to use an effective lea
for the computation of 32-rcx, we instead compute rcx-32 and
use an addition over a subtraction to carry out the comparison.

With some reshuffling of the code, this reduces the overhead of
this change to below noise level.

While we are at it, fix some inconsistent indentation in the
runt case.

Sponsored by:	The FreeBSD Foundation