lib/libc: add recallocarray()
Summary:
This function from OpenBSD is a hybrid of reallocarray() and calloc().
It reallocates an array, clearing any newly allocated items.

The source is taken from lib/libopenbsd, which now no longer has the
function.  It ultimately originates from OpenBSD.

A new version FBSD1.9 is added for this FreeBSD 16.0 symbol.

Reviewers: kib, imp

Subscribers: ziaee, bapt

Differential Revision: https://reviews.freebsd.org/D52863