sys/arm64: fix COMPAT_FREEBSD32 __syscall()
It seems like _QUAD_LOWWORD was incorrectly expanded into 1,
which is correct for big endian but not little endian.  This
means we always grab the padding word for the syscall number,
which is usually 0, causing SIGSYS to be delivered to the caller.
Reintroduce _QUAD_LOWWORD to fix the syscall.

PR:		290411
MFC after:	1 week
Reviewed by:	jrtc27
Approved by:	cognet, markj (mentor)
Fixes:		8c9c3144ccfa3061879b8cec015ee7d1010e4766
Differential Revision:	https://reviews.freebsd.org/D53250